+++ src/utils/swapCreator.ts
@@ -45,6 +45,7 @@ export type SwapBase = {
 export type SubmarineSwap = SwapBase &
     SubmarineCreatedResponse & {
         invoice: string;
+        preimage?: string;
         refundPrivateKey?: string;
     };
 
@@ -178,7 +179,9 @@ export const createChain = async (
         backend,
         assetSend,
         assetReceive,
-        Number(sendAmount),
+        sendAmount.isZero() || sendAmount.isNaN()
+            ? undefined
+            : Number(sendAmount),
         crypto.sha256(preimage).toString("hex"),
         claimKeys?.publicKey.toString("hex"),
         refundKeys?.publicKey.toString("hex"),